iotserver/requirements.txt
flask
kafka-python
python-dotenv